% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{ADMM}
\alias{ADMM}
\title{ADMM (Alternating Direction Method of Multipliers)}
\usage{
ADMM(tol_rel = 1e-05, tol_abs = 1e-04, alpha = 1.6, ...)
}
\arguments{
\item{tol_rel}{relative tolerance criterion for convergence}

\item{tol_abs}{absolute tolerance criterion for convergence}

\item{alpha}{over-regularization parameter, must be in \verb{(1.0, 1.8})}

\item{...}{ignored}
}
\value{
An object of class \code{c("ADMM", "Solver")}
}
\description{
ADMM (Alternating Direction Method of Multipliers)
}
