% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretGolem.R
\name{caretSlopeGolem}
\alias{caretSlopeGolem}
\title{Model objects for model tuning with caret}
\usage{
caretSlopeGolem()
}
\value{
A model description list to be used in the \code{method} argument
in \code{\link[caret:train]{caret::train()}}.
}
\description{
This function can be used in a call to \code{\link[caret:train]{caret::train()}} to enable
model tuning using caret. Note that this function does not properly work
with sparse feature matrices and standardization due to the way
resampling is implemented in caret. So for these cases, please
check out \code{\link[=trainGolem]{trainGolem()}} instead.
}
\seealso{
\code{\link[caret:train]{caret::train()}}, \code{\link[=trainGolem]{trainGolem()}}, \code{\link[=golem]{golem()}}
}
