% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance.Owl}
\alias{deviance.Owl}
\title{Model deviance}
\usage{
\method{deviance}{Owl}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'Owl'}.}

\item{...}{ignored}
}
\value{
For Gaussian models this is twice the residual sums of squares. For
all other models, two times the negative loglikelihood is returned.
}
\description{
Model deviance
}
\examples{
fit <- owl(heart$x, heart$y, family = "binomial")
deviance(fit)
}
