% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\docType{class}
\name{Diagnostics}
\alias{Diagnostics}
\title{Diagnostics}
\usage{
Diagnostics
}
\description{
This class encapsulates diagnostics from model fit with
\code{\link[=golem]{golem()}}.
}
\section{Fields}{

\describe{
\item{\code{data}}{contains a data frame with variables
\describe{
\item{\code{iteration}}{iteration}
\item{\code{time}}{time}
\item{\code{primal}}{primal objective}
\item{\code{dual}}{dual objective}
\item{\code{infeasibility}}{infeasibility}
\item{\code{penalty}}{the index along the regularization path}
}}
}}

\section{Methods}{


\describe{
\item{\code{plot(ind = "last", what = c("objectives", "infeasibility"))}}{
Plot objectives and infeasibility side-by-side using
trellis graphics.
\describe{
\item{\code{ind}}{
index of the fit for which diagnostics should be plotted
}
\item{\code{x_var}}{
what to place on the x axis. \code{iteration} plots each iteration, \code{time}
plots the wall-clock time.
}
\item{\code{y_var}}{
what to place on the y axis. \code{objectives} returns the
primal and dual objectives whereas \code{infeasibility} returns
the infeasibility metric.
}
\item{\code{\dots}}{
other arguments that will be used to modify the call to
\code{\link[lattice:xyplot]{lattice::xyplot()}}
}
}
}
}
}

