% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract_task.R
\name{extract_task}
\alias{extract_task}
\title{source and run task extraction script}
\usage{
extract_task(data, task)
}
\arguments{
\item{data}{all data split by subjects}

\item{task}{the name of the task, e.g. "fp"}
}
\value{
the results data frame from complete sessions of that task along with the date created and the "tags" that tell you the settings version.
}
\description{
source and run task extraction script
}
\examples{
extract_task(bySubject, "fp")
}

