% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{Fista}
\alias{Fista}
\title{FISTA solver}
\usage{
Fista(tol = 1e-05, max_passes = 10000, diagnostics = FALSE)
}
\arguments{
\item{tol}{tolerance threshold for convergence}

\item{max_passes}{maximum number of passes for the solver}

\item{diagnostics}{whether to save diagnostics (wall-clock time and loss)}
}
\value{
A set of parameters for the FISTA solver.
}
\description{
FISTA solver
}
