% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_for_csv.R
\name{format_for_csv}
\alias{format_for_csv}
\title{format_for_csv flattens any lists and reformats strings so Excel doesn't read them as dates}
\usage{
format_for_csv(raw.data)
}
\arguments{
\item{raw.data}{is a data frame of raw data for one task}
}
\value{
a data frame ready to be output as a csv file
}
\description{
format_for_csv flattens any lists and reformats strings so Excel doesn't read them as dates
}

