% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem-package.R
\docType{package}
\name{golem-package}
\alias{golem-package}
\title{Golem: Generalized Linear Models for Big Data}
\description{
Optimization routines for fitting generalized linear models to
big data.
}
\author{
Johan Larsson
}
\keyword{package}
