% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem.R
\name{golem}
\alias{golem}
\title{Golem}
\usage{
golem(x, y, family = c("gaussian", "binomial"), penalty = "SLOPE",
  solver = "auto", intercept = TRUE, lambda = "gaussian",
  sigma = 1, fdr = 0.2, standardize = c("features", "response",
  "both", "none"), max_passes = 1000, tol = 1e-04, ...)
}
\arguments{
\item{x}{asdf}

\item{...}{}
}
\value{
wef
}
\description{
Golem
}
\examples{
fit(3)
}
