% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{GroupSlope}
\alias{GroupSlope}
\title{Group SLOPE Penalty}
\usage{
GroupSlope(groups, lambda = c("corrected", "mean", "max"),
  sigma = NULL, fdr = 0.2, orthogonalize = TRUE)
}
\arguments{
\item{groups}{a vector of integers indicating which group each
feature belongs to}

\item{lambda}{penalty strength}

\item{sigma}{noise estimate}

\item{fdr}{target false discovery rate}

\item{orthogonalize}{whether to orthogonalize data matrix (within groups)}
}
\value{
A parameter pack for the Group SLOPE penalty.
}
\description{
Group SLOPE Penalty
}
