% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.owl.R
\name{import.owl}
\alias{import.owl}
\title{import OWL data, flatten and return a list of data split by subject}
\usage{
import.owl(choose = F)
}
\arguments{
\item{choose}{is a logical where TRUE means you have to select the file and FALSE means it searches for the "results.json" file, defaults to FALSE}
}
\value{
a list contatining all the data for each participant
}
\description{
import OWL data, flatten and return a list of data split by subject
}
\examples{
import.owl(choose = FALSE)
}

