% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaMax.R
\name{lambdaMax}
\alias{lambdaMax}
\alias{lambdaMax.Gaussian}
\alias{lambdaMax.Binomial}
\title{Smallest L1-regularization penalty where solution is completely sparse}
\usage{
lambdaMax(object, x, y, y_scale)

\method{lambdaMax}{Gaussian}(object, x, y, y_scale)

\method{lambdaMax}{Binomial}(object, x, y, y_scale)
}
\arguments{
\item{object}{an object of class \code{"Family"}}

\item{x}{feature matrix}

\item{y}{response}

\item{y_scale}{scale of response}
}
\value{
Lambda_max
}
\description{
Smallest L1-regularization penalty where solution is completely sparse
}
\keyword{internal}
