% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{Lasso}
\alias{Lasso}
\title{Lasso}
\usage{
Lasso(lambda = NULL, lambda_min_ratio = 1e-04, n_lambda = 100)
}
\arguments{
\item{lambda}{the regularization strength, which can either be a
user-supplied vector of (theoretical) any length, or \code{NULL}, in which
case golem automatically computes a sequence so that the first value
leads to the null (intercept-only) model.}

\item{lambda_min_ratio}{the lowest permissible value of the
lambda penalty}

\item{n_lambda}{the length of the \eqn{\lambda} sequence -- ignored if
a value is supplied to \code{lambda}.}
}
\value{
A parameter pack for the lasso penalty.
}
\description{
The lasso penalty penalized coefficients via the L1 norm, which induces
sparse solutions if the regularization strength (lambda) is sufficiently
strong.
}
