% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{linearPredictors}
\alias{linearPredictors}
\title{Return linear predictors}
\usage{
linearPredictors(object, x)
}
\arguments{
\item{object}{an ojbect of class \code{"Golem"}}

\item{x}{a feature matrix}
}
\value{
Linear predictors
}
\description{
Return linear predictors
}
\keyword{internal}
