% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl-package.R
\docType{package}
\name{owl-package}
\alias{owl}
\alias{owl-package}
\title{owl: Generalized Linear Models for Big Data}
\description{
Fast implementations of penalized Generalized Linear Models (GLM)
    suited to big data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/owl}
  \item Report bugs at \url{https://github.com/jolars/owl/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{mail@larssonjohan.com} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

Other contributors:
\itemize{
  \item Malgorzata Bogdan (code adapted from SLOPE and grpSLOPE) [contributor, copyright holder]
  \item Ewout van den Berg (code adapted from SLOPE) [contributor, copyright holder]
  \item Chiara Sabatti (code adapted from SLOPE) [contributor]
  \item Emmanuel Candes (code adapted from SLOPE) [contributor, copyright holder]
  \item Evan Patterson (code adapted from SLOPE) [contributor]
  \item Weijie Su (code adapted from SLOPE and grpSLOPE) [contributor, copyright holder]
  \item Damian Brzyski (code adapted from grpSLOPE) [contributor, copyright holder]
  \item Alexej Gossmann (code adapted from grpSLOPE) [contributor, copyright holder]
}

}
\keyword{internal}
