% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl_palettes.R
\docType{data}
\name{owl_palettes}
\alias{owl_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 5.}
\usage{
owl_palettes
}
\description{
Complete list of palettes
}
\keyword{datasets}

