% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Golem}
\alias{plot.Golem}
\title{Plot coefficients for a model fit by golem}
\usage{
\method{plot}{Golem}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"Golem"}}

\item{...}{parameters that will be used to modify the call to
\code{\link[lattice:xyplot]{lattice::xyplot()}}}
}
\value{
An object of class \code{"trellis"}, which will be plotted on the
current device unless stored in a variable.
}
\description{
Plot the model's coefficient along the regularization path.
}
\examples{
fit <- golem(heart$x, heart$y, penalty = "slope")
plot(fit)
}
\seealso{
\code{\link[lattice:xyplot]{lattice::xyplot()}}, \code{\link[=golem]{golem()}}, \code{\link[=plotDiagnostics]{plotDiagnostics()}}
}
