% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Owl}
\alias{plot.Owl}
\title{Plot coefficients}
\usage{
\method{plot}{Owl}(x, intercept = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{"Owl"}}

\item{intercept}{whether to plot the intercept}

\item{...}{parameters that will be used to modify the call to
\code{\link[lattice:xyplot]{lattice::xyplot()}}}
}
\value{
An object of class \code{"trellis"}, which will be plotted on the
current device unless stored in a variable.
}
\description{
Plot the model's coefficient along the regularization path.
}
\examples{
fit <- owl(heart$x, heart$y)
plot(fit)
}
\seealso{
\code{\link[lattice:xyplot]{lattice::xyplot()}}, \code{\link[=owl]{owl()}}, \code{\link[=plotDiagnostics]{plotDiagnostics()}}
}
