% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict,Golem-method}
\alias{predict,Golem-method}
\title{Make Predictions from Golem Models}
\usage{
\S4method{predict}{Golem}(object, newdata, type = c("link", "response",
  "class"), ...)
}
\arguments{
\item{object}{an object of class "Golem"}

\item{newdata}{new data to make predictions for}

\item{type}{type of predictions. \code{"link"} returns the linear predictors,
\code{"response"} returns the response after applying the link functions,
and \code{"class"} returns the predicted class (for the binomial family).}

\item{...}{ignored.}
}
\value{
A vector of predictions.
}
\description{
Make Predictions from Golem Models
}
\examples{
x <- as.matrix(with(Puromycin, cbind(conc, rate)))
y <- Puromycin$state

fit <- golem(x, y, family = "binomial")
predict(fit, x, "class")
}
