% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\alias{score.GolemGaussian}
\alias{score.GolemBinomial}
\title{Return the loss of a golem model}
\usage{
score(object, x, y, measure)

\method{score}{GolemGaussian}(object, x, y, measure = c("deviance",
  "mse", "mae"))

\method{score}{GolemBinomial}(object, x, y, measure = c("deviance",
  "mse", "mae", "accuracy", "auc"))
}
\arguments{
\item{object}{an object of class \code{"Golem"}}

\item{x}{feature matrix}

\item{y}{response}

\item{measure}{type of target measure. The default, \code{"deviance"},
returns mean squared error for Gaussian models and deviance for
Binomial models. \code{"mse"} returns mean squared error. \code{"mae"} returns
mean absolute error, \code{"accuracy"} returns classification rate accuracy,
and \code{"auc"} returns area under the ROC curve.}
}
\value{
The measure along the regularization path depending on the
value in \code{measure}.
}
\description{
This function is a unified interface to return various types of loss for a
model fit with \code{\link[=golem]{golem()}}.
}
\examples{
x <- subset(infert, select = c("induced", "age", "pooled.stratum"))
y <- infert$case

fit <- golem(x, y, family = "binomial")
score(fit, x, y, measure = "auc")
}
