% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenFeatures.R
\name{screenFeatures}
\alias{screenFeatures}
\title{Screen features}
\usage{
screenFeatures(
  family,
  penalty,
  x,
  y,
  lambda,
  lambda_prev,
  beta_prev,
  intercept_prev,
  gradient_prev,
  method = c("none", "strong", "safe")
)
}
\arguments{
\item{family}{family}

\item{penalty}{penalty}

\item{x}{features}

\item{y}{response}

\item{lambda}{regularization sequence of current fit}

\item{lambda_prev}{regularization sequence of previous fit}

\item{beta_prev}{coefficients from previous fit}

\item{intercept_prev}{intercept from previous fit}

\item{gradient_prev}{gradient from previous fit}

\item{method}{type of screening}
}
\value{
A logical vector indicating whether to drop a feature or not
}
\description{
Screen features
}
