% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeCaret.R
\name{slopeCaret}
\alias{slopeCaret}
\title{Model objects for model tuning with caret}
\usage{
slopeCaret()
}
\value{
Please see \code{\link[caret:train]{caret::train()}}
}
\description{
This function can be used in a call to \code{\link[caret:train]{caret::train()}} to enable
model tuning using caret. This function is currently not exported
pending investigation.
}
\keyword{internal}
