% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split.task.results.R
\name{split.task.results}
\alias{split.task.results}
\title{split one participant's results by task}
\usage{
\method{split}{task.results}(df, task = task)
}
\arguments{
\item{df}{data frame containing one subject's data to be split}

\item{task}{the name of the task, e.g. "fp"}
}
\value{
the results data frame from complete sessions of that task along with the date created and the "tags" that tell you the settings version.
}
\description{
split one participant's results by task
}
\examples{
split.task.results(df = bySubject[[1]], task = "fp")
}

