% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task.status.R
\name{task.status}
\alias{task.status}
\title{check the status of a task}
\usage{
task.status(task, list = bySubject)
}
\arguments{
\item{task}{is the task to be summarised}

\item{data}{is the list of subject data (defaults to buSubject)}
}
\value{
a data frame of information about the number of participants who have completed, abandoned or not started the task and
at which trial they abandoned
}
\description{
check the status of a task
}
\examples{
task.status("fp", bySubject)
}

