% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{trainGolem}
\alias{trainGolem}
\title{Train (tune) model parameter for a golem model}
\usage{
trainGolem(x, y, groups = NULL, method = "cv", number = 10,
  repeats = 1, measure = c("deviance", "mse", "mae", "misclass_error",
  "auc"), n_cores = 1, ...)
}
\arguments{
\item{x}{the feature matrix, which can be either a dense
matrix of the standard \emph{matrix} class, or a sparse matrix
inheriting from \link[Matrix:sparseMatrix]{Matrix::sparseMatrix} Data frames will
be converted to matrices internally.}

\item{y}{the response. For Gaussian models this must be numeric; for
binomial models, it can be a factor.}

\item{groups}{a vector of integers giving the group membership of each
feature (only applies to Group SLOPE)}

\item{method}{type of model tuning method}

\item{number}{number of folds (cross-validation)}

\item{repeats}{number of repeats for each fold (for repeated \emph{k}-fold
cross validation)}

\item{measure}{measure to try to optimize}

\item{n_cores}{number of cores to use for parallel processing}

\item{...}{other arguments to pass on to \code{\link[=golem]{golem()}}}
}
\value{
An object of class \code{"TrainedGolem"}
}
\description{
Train (tune) model parameter for a golem model
}
\seealso{
\link[parallel:parallel]{parallel::parallel}
}
