% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.summaries.R
\name{variable.summaries}
\alias{variable.summaries}
\title{get summaries of each of the variables for a given task}
\usage{
variable.summaries(task, data.sum)
}
\arguments{
\item{task}{is the task to be summarised}

\item{data.sum}{is the summary data to be summarised further}
}
\value{
a data frame with summaries of each of the variables in the data
}
\description{
get summaries of each of the variables for a given task
}
\examples{
variable.summaries("fp", summary_data)
}

