% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{what}
\alias{what}
\title{Fit a Model with sgdnet}
\usage{
what(x, y)
}
\arguments{
\item{x}{a double}

\item{y}{a double}
}
\value{
Something
}
\description{
This main use of this function is calling the templated SetupSgdnet()
so that the dense and sparse implementations are compiled and
called appropriately. The control parameters in \code{control} are just
passed along.
}
