% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.Golem}
\alias{coef.Golem}
\title{Return Coefficients from a Golem Model}
\usage{
\method{coef}{Golem}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"Golem"}}

\item{...}{parameters passed onto \code{\link[stats:coef]{stats::coef()}}}
}
\value{
The coefficients, including the intercept if it was fit.
}
\description{
Returns the coefficients for a model fit via \code{\link[golem:golem]{golem::golem()}}.
Unlike retrieving the coefficients using simple indexing, such as
\code{fit$coefficients}, this function attempts to reduce the dimensions of
the coefficients---that are otherwise always a three-dimensional array---by
calling \code{\link[base:drop]{base::drop()}}.
}
