% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.Golem}
\alias{coef.Golem}
\title{Obtain Coefficients from Model fit by Golem}
\usage{
\method{coef}{Golem}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'Golem'}.}

\item{...}{ignored}
}
\value{
Coefficients from the model after having dropped extraneous
dimensions by calling drop.
}
\description{
This function is equivalent to calling \code{$coef()} on a model fit by
\code{\link[=golem]{golem()}}.
}
