% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\docType{class}
\name{Diagnostics}
\alias{Diagnostics}
\title{Diagnostics}
\usage{
Diagnostics
}
\description{
This class encapsulates diagnostics from model fit with
\code{\link[=golem]{golem()}}.
}
\section{Fields}{

\describe{
\item{\code{data}}{contains a data frame with variables
\describe{
\item{\code{time}}{time}
\item{\code{primal}}{primal objective}
\item{\code{dual}}{dual objective}
\item{\code{infeasibility}}{infeasibility}
\item{\code{penalty}}{the index along the regularization path}
}}
}}

\section{Methods}{


\describe{
\item{\code{plot()}}{
Plot objectives and infeasibility side-by-side using
trellis graphics.
}
}
}

