% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{FISTA}
\alias{FISTA}
\title{FISTA (Fast Iterative Shrinking--Thresholding Algorithm)}
\usage{
FISTA(tol_rel_gap = 1e-05, tol_infeas = 1e-04)
}
\arguments{
\item{tol_rel_gap}{relative tolerance criterion for the duality gap}

\item{tol_infeas}{tolerance criterion for infeasibility of the
dual objective}
}
\value{
An object of class \code{c("Fista", "Solver")}
}
\description{
FISTA is a first-order gradient method.
}
