% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{fista}
\alias{fista}
\title{FISTA solver}
\usage{
fista(tol = 1e-06, max_passes = 10000)
}
\arguments{
\item{tol}{tolerance threshold for convergence}

\item{max_passes}{maximum number of passes for the solver}
}
\value{
A set of parameters for the FISTA solver.
}
\description{
FISTA solver
}
