% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{FISTA}
\alias{FISTA}
\title{FISTA (Fast Iterative Shrinking--Thresholding Algorithm)}
\usage{
FISTA(
  tol_rel_gap = 1e-05,
  tol_infeas = 1e-04,
  line_search = FALSE,
  line_search_frequency = 1,
  ...
)
}
\arguments{
\item{tol_rel_gap}{relative tolerance criterion for the duality gap}

\item{tol_infeas}{tolerance criterion for infeasibility of the
dual objective}

\item{line_search}{whether to perform line search}

\item{line_search_frequency}{at which iterations to perform line search}

\item{...}{ignored}
}
\value{
An object of class \code{c("Fista", "Solver")}
}
\description{
FISTA is a first-order gradient method.
}
