% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl-package.R
\docType{package}
\name{owl-package}
\alias{owl-package}
\alias{_PACKAGE}
\title{owl: Generalized Linear Models Regularized with the Sorted L1-Norm}
\description{
Efficient implementations for Sorted L-One Penalized Estimation
    (SLOPE): generalized linear models regularized with the 
    sorted L1-norm (Bogdan et al. (2015) <doi:10/gfgwzt>)
    or, equivalently, ordered weighted L1-norm (OWL). Supported models include
    ordinary least-squares regression, binomial regression, multinomial
    regression, and Poisson regression. Both dense and sparse 
    predictor matrices are supported. In addition, the package features
    predictor screening rules that enable fast and efficient solutions to 
    high-dimensional problems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/owl}
  \item \url{https://jolars.github.io/owl}
  \item Report bugs at \url{https://github.com/jolars/owl/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johan.larsson@stat.lu.se} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

Authors:
\itemize{
  \item Jonas Wallin \email{jonas.wallin@stat.lu.se} (\href{https://orcid.org/0000-0003-0381-6593}{ORCID})
}

Other contributors:
\itemize{
  \item Malgorzata Bogdan (code adapted from 'SLOPE') [contributor]
  \item Ewout van den Berg (code adapted from 'SLOPE') [contributor]
  \item Chiara Sabatti (code adapted from 'SLOPE') [contributor]
  \item Emmanuel Candes (code adapted from 'SLOPE') [contributor]
  \item Evan Patterson (code adapted from 'SLOPE') [contributor]
  \item Weijie Su (code adapted from 'SLOPE') [contributor]
  \item Jerome Friedman (code adapted from 'glmnet') [contributor]
  \item Trevor Hastie (code adapted from 'glmnet') [contributor]
  \item Rob Tibshirani (code adapted from 'glmnet') [contributor]
  \item Balasubramanian Narasimhan (code adapted from 'glmnet') [contributor]
  \item Noah Simon (code adapted from 'glmnet') [contributor]
  \item Junyang Qian (code adapted from 'glmnet') [contributor]
}

}
\keyword{internal}
