% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.Golem}
\alias{predict.Golem}
\alias{predict.GolemGaussian}
\alias{predict.GolemBinomial}
\title{Title}
\usage{
\method{predict}{Golem}(object, newdata = NULL, type = c("link",
  "response", "terms"), ...)

\method{predict}{GolemGaussian}(object, newdata = NULL,
  type = c("link", "response", "terms"), ...)

\method{predict}{GolemBinomial}(object, newdata = NULL,
  type = c("link", "response", "terms", "class"), ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
