% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.Golem}
\alias{predict.Golem}
\alias{predict.GolemGaussian}
\alias{predict.GolemBinomial}
\title{Generate predictions from golem models}
\usage{
\method{predict}{Golem}(object, x, lambda = NULL, sigma = NULL,
  type = "link", exact = FALSE, simplify = TRUE, ...)

\method{predict}{GolemGaussian}(object, x, lambda = NULL, sigma = NULL,
  type = c("link", "response"), exact = FALSE, simplify = TRUE, ...)

\method{predict}{GolemBinomial}(object, x, lambda = NULL, sigma = NULL,
  type = c("link", "response", "class"), exact = FALSE,
  simplify = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{"Golem"}, typically the result of
a call to \code{\link[=golem]{golem()}}}

\item{x}{new data}

\item{lambda}{penalty parameter for Lasso-type models; if \code{NULL}, the
values used in the original fit will be used}

\item{sigma}{penalty parameter for SLOPE models; if \code{NULL}, the
values used in the original fit will be used}

\item{type}{type of prediction; \code{"link"} returns the linear predictors,
\code{"response"} returns the result of applying the link function,
and \code{"class"} returns class predictions.}

\item{exact}{if \code{TRUE} and the given parameter values differ from those in
the original fit, the model will be refit by calling \code{\link[stats:update]{stats::update()}} on
the object with the new parameters. If \code{FALSE}, the predicted values
will be based on interpolated coefficients from the original
penalty path.}

\item{simplify}{if \code{TRUE}, \code{\link[base:drop]{base::drop()}} will be called before returning
the coefficients to drop extraneous dimensions}

\item{...}{ignored and only here for method consistency}
}
\value{
Predictions from the model with scale determined by \code{type}.
}
\description{
Return predictionsfrom models fit by \code{\link[=golem]{golem()}}.
}
\examples{
fit <- with(mtcars, golem(cbind(mpg, hp), vs, family = "binomial"))
predict(fit, with(mtcars, cbind(mpg, hp)), type = "class")

}
\seealso{
\code{\link[stats:predict]{stats::predict()}}, \code{\link[stats:predict.glm]{stats::predict.glm()}}
}
