% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.Golem}
\alias{predict.Golem}
\alias{predict.GolemGaussian}
\alias{predict.GolemBinomial}
\title{Generate predictions from golem models}
\usage{
\method{predict}{Golem}(object, x, type, ...)

\method{predict}{GolemGaussian}(object, x, type = c("link", "response"),
  ...)

\method{predict}{GolemBinomial}(object, x, type = c("link", "response",
  "class"), ...)
}
\arguments{
\item{object}{an object of class \code{"Golem"}, typically the result of
a call to \code{\link[=golem]{golem()}}}

\item{x}{new data}

\item{type}{type of prediction; \code{"link"} returns the linear predictors,
\code{"response"} returns the result of applying the link function,
and \code{"class"} returns class predictions.}

\item{...}{ignored and only here for method consistency}
}
\value{
Predictions from the model with scale determined by \code{type}.
}
\description{
Return predictions from models fit by \code{\link[=golem]{golem()}} based on
new data.
}
\examples{
fit <- with(mtcars, golem(cbind(mpg, hp), vs, family = "binomial"))
predict(fit, with(mtcars, cbind(mpg, hp)), type = "class")

}
\seealso{
\code{\link[stats:predict]{stats::predict()}}, \code{\link[stats:predict.glm]{stats::predict.glm()}}
}
