% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Owl}
\alias{print.Owl}
\alias{print.TrainedOwl}
\title{Print results from owl fit}
\usage{
\method{print}{Owl}(x, ...)

\method{print}{TrainedOwl}(x, ...)
}
\arguments{
\item{x}{an object of class \code{'Owl'} or \code{'TrainedOwl'}}

\item{...}{other arguments passed to \code{\link[=print]{print()}}}
}
\value{
Prints output on the screen
}
\description{
Print results from owl fit
}
\examples{
fit <- owl(wine$x, wine$y, family = "multinomial")
print(fit, digits = 1)

}
