% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{Slope}
\alias{Slope}
\title{Sorted L-One Penalized Estimation (SLOPE)}
\usage{
Slope(lambda = c("gaussian", "bhq"), sigma = NULL, fdr = 0.2)
}
\arguments{
\item{lambda}{penalty strength}

\item{sigma}{estimate of signal noise}

\item{fdr}{target for false discovery rate (FDR)}
}
\value{
A parameter pack for the SLOPE penalty.
}
\description{
Sorted L-One Penalized Estimation (SLOPE)
}
